+++ cli/commands/mongodb.go
@@ -21,7 +21,7 @@ var mongoDbCmd = &cobra.Command{
   url      = "%s"
 }
 `, databaseURL)
-		prismaSchema, graphqlSDL, dmmf, err := database.IntrospectPrismaDatabase(introspectionSchema, log)
+		prismaSchema, graphqlSDL, dmmf, err := database.IntrospectPrismaDatabase(introspectionSchema, WunderGraphDir, log)
 		if err != nil {
 			return err
 		}