+++ packages/sdk/src/configure/index.ts
@@ -1032,6 +1032,7 @@ const mapDataSource = (source: DataSource): DataSourceConfiguration => {
 				subscription: {
 					enabled: graphql.Subscription.Enabled,
 					url: graphql.Subscription.URL,
+					useSSE: graphql.Subscription.UseSSE,
 				},
 				upstreamSchema: graphql.UpstreamSchema,
 			};