+++ packages/swr/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.3.4](https://github.com/wundergraph/wundergraph/compare/@wundergraph/swr@0.3.3...@wundergraph/swr@0.3.4) (2022-09-29)
+
+**Note:** Version bump only for package @wundergraph/swr
+
+## [0.3.3](https://github.com/wundergraph/wundergraph/compare/@wundergraph/swr@0.3.2...@wundergraph/swr@0.3.3) (2022-09-29)
+
+**Note:** Version bump only for package @wundergraph/swr
+
 ## [0.3.2](https://github.com/wundergraph/wundergraph/compare/@wundergraph/swr@0.3.1...@wundergraph/swr@0.3.2) (2022-09-26)
 
 **Note:** Version bump only for package @wundergraph/swr